+++ src/PDFs/combine/CMakeLists.txt
@@ -1,16 +1,18 @@
-goofit_add_library(PdfCombine
+# This is only needed for IDE header display
+file(GLOB_RECURSE GOOFIT_COMBINE_HEADERS
+    "${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/*.h"
+    )
+
+goofit_add_library(PDFCombine
     AddPdf.cu
     CompositePdf.cu
     ConvolutionPdf.cu
     EventWeightedAddPdf.cu
     MappedPdf.cu
     ProdPdf.cu
-    ${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/AddPdf.h
-    ${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/CompositePdf.h
-    ${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/ConvolutionPdf.h
-    ${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/EventWeightedAddPdf.h
-    ${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/MappedPdf.h
-    ${PROJECT_SOURCE_DIR}/include/goofit/PDFs/combine/ProdPdf.h
+    ${GOOFIT_COMBINE_HEADERS}
 )
 
-target_link_libraries(PdfCombine PUBLIC PDFCore)
+set_target_properties(PDFCombine PROPERTIES FOLDER "pdfs")
+
+target_link_libraries(PDFCombine PUBLIC PDFCore)
